feat: support stats child assets
